Moving logic to logic.js
