+++ src/logic.js
@@ -1,4 +1,12 @@
+const listCountElement = document.querySelector('[data-list-count]');
+const listsContainer = document.querySelector('[data-lists]');
 
+  
+  const LOCAL_STORAGE_LIST_KEY = 'task.lists';
+  const LOCAL_STORAGE_SELECTED_LIST_ID_KEY = 'task.selectedListsId';
+
+  let lists = JSON.parse(localStorage.getItem(LOCAL_STORAGE_LIST_KEY)) || [];
+  let selectedListId = localStorage.getItem(LOCAL_STORAGE_SELECTED_LIST_ID_KEY);
   function createList(name) {
     return { id: Date.now().toString(), name, tasks: [] };
   }
@@ -13,20 +21,29 @@
     }
   }
 
-  export function render() {
-    clearElement(listsContainer);
-    renderLists();
-
-    const selectedList = lists.find((list) => list.id === selectedListId);
-    if (selectedListId == null) {
-      listDisplayContainer.style.display = 'none';
-    } else {
-      listDisplayContainer.style.display = '';
-      listTitleElement.innerText = selectedList.name;
-      renderTaskCount(selectedList);
-      clearElement(tasksContainer);
-      renderTasks(selectedList);
-    }
+  function save() {
+    localStorage.setItem(LOCAL_STORAGE_LIST_KEY, JSON.stringify(lists));
+    localStorage.setItem(LOCAL_STORAGE_SELECTED_LIST_ID_KEY, selectedListId);
   }
 
-  export { createList, createTask, clearElement}
+  function renderTaskCount(selectedListId) {
+    const incompleteTaskCount = selectedListId.tasks.filter((task) => !task.complete).length;
+    const taskString = incompleteTaskCount === 1 ? 'task' : 'tasks';
+    listCountElement.innerText = `${incompleteTaskCount} ${taskString} remaining`;
+  }
+
+  function renderLists() {
+    lists.forEach((list) => {
+      const listElement = document.createElement('li');
+      listElement.dataset.listId = list.id;
+      listElement.classList.add('list-name');
+      listElement.innerText = list.name;
+      if (list.id === selectedListId) {
+        listElement.classList.add('active-list');
+      }
+      listsContainer.appendChild(listElement);
+    });
+  }
+
+
+  export { renderLists, createList, createTask, clearElement, save, renderTaskCount}
